requirements.txt
flask
pillow
requests